restructure docs
